[refactor] Move `UserSettingsFeatureTests` to single file
